backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
anthropic>=0.52.0
numpy>=2.0.0
Pillow>=11.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
opencv-python>=4.9.0
plyfile>=1.0.0
tqdm>=4.66.0
gsplat>=1.5.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
